+++ src/schema/__tests__/RefraxTreeNode.spec.js
@@ -8,6 +8,10 @@
 const chai = require('chai');
 const RefraxTreeNode = require('RefraxTreeNode');
 const expect = chai.expect;
+const RefraxConstants = require('RefraxConstants');
+const CLASSIFY_COLLECTION = RefraxConstants.classify.collection;
+const CLASSIFY_RESOURCE = RefraxConstants.classify.resource;
+
 
 
 /* eslint-disable no-new */
@@ -15,40 +19,40 @@ describe('RefraxTreeNode', function() {
   describe('instantiation', function() {
     it('should only accept object params type', function() {
       expect(function() {
-        new RefraxTreeNode(123);
+        new RefraxTreeNode(CLASSIFY_RESOURCE, 123);
       }).to.throw(Error, 'pass an invalid definition of type');
 
       expect(function() {
-        new RefraxTreeNode('abc');
+        new RefraxTreeNode(CLASSIFY_RESOURCE, 'abc');
       }).to.throw(Error, 'pass an invalid definition of type');
     });
 
     it('should throw an error with an invalid option', function() {
       expect(function() {
-        new RefraxTreeNode({barfoo: 123});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {barfoo: 123});
       }).to.throw(Error, 'Invalid definition option');
     });
 
     it('should throw an error on invalid option values', function() {
       expect(function() {
-        new RefraxTreeNode({partial: 123});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {partial: 123});
       }).to.throw(Error);
 
       expect(function() {
-        new RefraxTreeNode({fragments: 123});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {fragments: 123});
       }).to.throw(Error);
 
       expect(function() {
-        new RefraxTreeNode({uri: 123});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {uri: 123});
       }).to.throw(Error);
     });
 
     it('should not throw an error on valid option values', function() {
       expect(function() {
-        new RefraxTreeNode({uri: '/foo'});
-        new RefraxTreeNode({partial: 'minimal'});
-        new RefraxTreeNode({paramId: 'fooId'});
-        new RefraxTreeNode({fragments: ['default', 'full']});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {uri: '/foo'});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {partial: 'minimal'});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {paramId: 'fooId'});
+        new RefraxTreeNode(CLASSIFY_RESOURCE, {fragments: ['default', 'full']});
       }).to.not.throw(Error);
     });
   });